[FULL] Add workaround  in jshost to not throw while release timeZone objects
My changes in Intl that added support for timeZoneOnCalendar exposed an interim issue.
We hold windows.globalization's timeZoneOnCalendar which holds to mrmcorer.dll. When we
do CoUninitialize() in jshost, the mrmcorer.dll is unloaded. Because of this while releasing
timeZoneOnCalendar we get AV because it tries to release freed contents from mrmcorer.dll.
As a workaround, we manually release the timeZoneOnCalendar before calling CoUninitialize().
This problem should only be present in jshost.exe and IE handles the ordering of CoUninitialize()
after DllUnload(). Thanks to Yong for helping find this and suggesting a fix.