Partial clean up warning 4267
C4267 is a warning flagged by binscope as SDL-required
C4267 is 'var' : conversion from 'size_t' to 'type', possible loss of data.
Truncation can lead to buffer overrun around allocation or length calculation.
So it is good to clean up.

This change clean up warnings in some of the files.  But there are too many warnings to clean up all at once.
So the warning is disabled for files that still having the warning instead of disabling the warning project wide.