Baseline changes accompanying the removal of anonymous function name
This change contains the baseline changes from removing the name from
anonymous functions. It coorespnds with a change in core that does two
things. First it creates an anonymous function typehandler that does not
use a deffered typehadnler that filters for a name property. Second it
removes the name property from Function.prototype.name b\c of a spec
change from April to June 2015