Modify the path to lower text for jhost.
This is to adjust with ch.exe. As the both jshost and ch lowers the fullpath and pass them to chakra/chakracore. However the jshost has different mechanism to print the stack. In order to fix this I have just lowered the filename text during the printing  stack code path.
     Updated baseline