Update version and file description in resources
Add branch and commit hash info in the VSO CI build, for both core and private.
Also set the version ChakraCore to start with 1.0 for now.
Build number will be the pushed number once the VSO CI is configured with a prebuild script
to set ChakraVersionBuildNumber in the environment.

Chakra.dll will remain the same as the IE version number for now.