Fix security bug OS 5120001 - Stack Buffer Overflow in SpreadArgs
The arguments object has a configurable length property and so can be
changed to be a user script getter property. The code flow for
CallSpreadFunction is doing two gets of the spread argument's length,
and so in the case of an arguments object with accessor lenght, calls
user code twice.

The first call is to figure out how much buffer to allocate on the stack
for the arguments and the second call is to figure out how many elements
to copy from the spread argument into the buffer.

With user script being called twice for both requests of the length,
malicious code had the opportunity to return a small value first, and
then a large value second, causing SpreadArgs to write past the end of
the destination buffer and overwrite stack memory.

The bug includes a repro that shows the stack pointer to bar for the
call is overwritten by the attack with a pointer to the evilCat
function. Write a little further up the stack and the return address
can be overwritten with anything the user wants.

This attack is now protected against by including bounds checking when
copying to the destination buffer.

This bug also revealed some incorrect optimizations in the spread code.
First the arguments object cannot be optimized away from the @@iterator
pattern because of the configurable/writable length property and the
ability for it to have accessor indexed properties. Second we cannot
optimize arrays when the %ArrayIteratorPrototype%.next property is no
longer pointing to the corresponding built-in function.

The former issue I have corrected by disallowing arguments objects from
being optimized by wrapping them also in SpreadArgument wrapper class
during the LdCustomSpreadIteratorList bytecode op. The latter I have
not addressed here because it requires getting the value stored in the
next property of the %ArrayIteratorPrototype% built-in object without
invoking the property if it is a getter and I don't believe we have a
way to do that at the moment. Instead I have opened bug 5204357 to
track this. I have included unit tests for these two issues and for the
ES5Array equivalent of the arguments object issue, although that case
already passed because ES5Array objects were already being wrapped in
the SpreadArgument helper class.