Remove usage of RtlGetDeviceFamilyInfoEnum in rl
Remove RltGetDeviceFamilyInfoEnum API usage in rl.exe source.  It is
undocumented API and cannot be part of OSS code.  Appears it was added
to support running rl.exe on Apollo mobile machines.  We no longer run
rl in this mode.  Removing all related code for now.  We can re-enable
such support in the future if we need to when it comes up.