78859 (OSG VSO 5119253) NewTree Completion#1: Provide support from Chakra to catch unexpected script re-entrancy
[NOTE: this change is already checked into RS1_ONECORE_MQEDGE1_STAGE]

The NoScriptScope functionality needed by Trident for TreeWriter requires a
component in the Chakra engine. This adds the SetNoScriptScope and
IsNoScriptScope APIs to the EdgeJsStatic exports and the associated changes and
tests within the Chakra code base for this functionality.

In particular, the API allows the caller to set the NoScriptScope flag within
the JavascriptThreadContext, and this flag will be honored by failing fast on
EnterScript if it is set, to prevent execution of scripts within a
NoScriptScope.

This functionality enables:
* IECAT 78296 (http://iecat/viewchange.aspx?id=78296)
* OSG VSO 4877539 (https://microsoft.visualstudio.com/DefaultCollection/OS/_workitems#_a=edit&id=4877539)