Destructuring fix by saving AST information.
In order to satisfy the current grammar as destructuring grammar we parse the text again (with destructuring grammar). However we used to save only scope information. There are others which we need to save and restore back as we will going to invalid them in this reparsing process. Fixed that and added a unittest for that.