Object literal shorthand issue.
Object literal shorthand issue.

In the ParseMemberList we use different way of adding pidref when we determine it is shorthand syntax. Ideally {a} should behave same as {a:a} in terms of stack pid reference look-up. In the case of {a:a} we actually do the PushPidRef however in the case of shorthand we do the FindOrAddPidRef(). The latter actually mess the up pidrefstack linked list.
Fix this by doing the PushPidRef.
Added the unittest.