Add timeout function to RL test runner
Add timeout function to RL test runner

Default timeout per test is 60 seconds.  Can be overridden by specifying
a timeout in seconds with the `<timeout>` tag in the `rlexe.xml` config
file.

When a test times out the output will look like the following:

```
3>+++ expectedfail (%REGRESS%\mediumrunning.js () exe) +++
3>Running '%REGRESS%\runjsfile.cmd %REGRESS%\mediumrunning.js   -DumpOnCrash -WERExc...
3>SUCCESS: The process with PID 26204 (child process of PID 35960) has been terminated.
3>SUCCESS: The process with PID 35960 (child process of PID 7200) has been terminated.
3>ERROR: Test failed to run correctly: timed out after 1 second ():
3>    %REGRESS%\runjsfile.cmd %REGRESS%\mediumrunning.js   -DumpOnCrash -WERExceptio...
3>ERROR: name of output file: W:\ch1\core\test\rltimeout\expectedfail\testout4; size...
3>ERROR: bad output file follows ============
3>
3>[W:\ch1 x64debug] W:\ch1\core\test\rltimeout\expectedfail>ch.exe W:\ch1\core\test\...
3>ERROR: end of bad output file  ============
3>RL: Test elapsed time (expectedfail, UnnamedTest6): 00:01
```

In this case the timeout was specified at one second which can be seen
on the first `ERROR:` line.  The log still shows the output that the
test produced up to the timeout point to aid in diagnosis.

`taskkill.exe` is used to kill the test cmd process and its children.
It prints out `SUCCESS: ...` for each process it terminates.

There is no switch to override the default globally provided in this
change.  It can be added later if the need arises.