Fix for final type + ctor cache + inlining
[Core Commit Message]
Fix for a bug involving a collision of final type opt, constructor cache, and inlining. An inlined constructor can advance the constructed object's type past the helper call that updates the constructor cache. If that happens, and if a later constructor bails out in the middle, the object's type may contain properties that it doesn't really have. Simple fix: don't let GlobOpt push a final type up past the UpdateNewScObjectCache instruction. (We'll revisit the way final type and ctor cache interact later.)