Fix LdHeapArguments with rest parameter as the only formal [Update submodule only]
[Core Commit Message]
The LdHeapArguments implementation in the interpreter/JIT relies on the formal argument count to tell it whether to pass a real frame object or JS null to the helper. But this doesn't work in the case where the only formal is a rest parameter. I don't know whether there are (or will be) other special cases. Instead, just look and see whether we've loaded a property ID array. If we have, then the helper needs the frame object. Also clean up some cruft from the last byte code hardening change.