Fix non-debug bail out at offset 0 [Update submodule only]
[Core Commit Message]
We've been using a bail out offset of zero to indicate that no closure init has been done in the interpreter. But it's possible for a jitted function to bail out at offset zero after closure init. Detect this case in bailout and notify the interpreter instance through a flag so it has a reliable way of distinguishing the cases.