Fix debugger bug introduced in for loop semantics change
I introduced a bug in commit 19e8ed5e7183615f91709bffa884362d01aa7dfe
where if a for-in or for-of loop didn't have let or const loop variables
then any let or const loop variables in the scope containing the for
loop were marked as being in the TDZ via the call to
RecordForInOrOfCollectionScope().  This function should only been called
if the for loop block has let or const variable bindings.