Add daily build tasks.
We are adding some daily build tasks which will build in a configuration not usually used for our PR and per-commit builds.

We don't want to add daily builds of the same configurations we already use for each PR and each commit, as they would be redundant. However there are some configurations that are worth testing at least on a daily basis: {x86, x64} x {debug, test, release} x {Windows 7} x {MsBuild 12 from VS 2013 aka Dev12}.

For these configurations we add a `@daily` task with a trigger to allow a commenter to request `@dotnet-bot` to run the daily build configurations on any PR if desired, in the same way that `test this please` can be used to trigger all of the usual tests.