Implement [no LineTerminator here] grammar rules for async functions
The async pseudokeyword is only recognized when it is not followed by a line terminator.  E.g. the following code does not parse an async function:

```javascript
var a = async
    function f() { }
```

Instead it parses a var declaration whose initializer expression is `async` and then a function declaration named `f` due to automatic semicolon insertion.