Empty pattern for destructuring case.
Empty pattern for destructuring case.

Catch can take empty destructuring pattern, in that case we don't need to instantiate the scope as there is no need to create slot array or scope object because there is no catch param. Add a check as count being nonzero.
Another place (parse.cpp) where the m_currentNodeFunc is null because that is deferred parse case. Move the code under buildAST flag (as it was done other places). Added tests