Memset on typed specialized values
Sometimes, when type specializing the original instruction is lost.
ie: `s3.var          =  LdC_A_I4       0 (0x0).i32` is type specialized to `int32`, but the dst is used as a float.
If it's the only use, then `s3.var` is removed.
Previously memset was trying to get `s3.var` from the float type specialized version.
However, it is not live anymore.

This change makes memset use directly the value the `StElemI_A` is using (ie: the float typed specialized register)
And will call ToVar on it.
A better fix would be to try to use the var version first if it is available, but might introduce unexpected bugs.

I also added several test cases to cover this issue and more, along with some tests with simds.