Fix static semantic errors for arrow and async function parameter lists
Arrow function parameter lists were being parsed as always strict.  This is incorrect.  The spec only says that they should have no duplicate names via the rules for the `StrictFormalParameters` grammar production. `arguments` and `eval` are allowed to be parameter names of arrow functions in non-strict mode.

Fixing that required adding a "force pid" version of the scanner's `SeekTo` method for rewinding.  The new method is called `SeekToForcingPid`.

Async function parameter lists were not applying strict mode rules when the async function itself turned on strict mode.  This was due to the inner generator function becoming strict but not propagating the strictness out to the enclosing function that has the user parameters.

The fix for that was to simply propagate the strictness of the inner generator function to the enclosing function.