Define a SimpleTypeHandler for functions with length and name properties
When creating a function's type without prototype, we instantiate a
SimpleTypeHandler, with a max property size of 1, with just the length
property. Later, when we add the name
property we end up having to convert the simple type handler to simple
dictionary type handler.

This change prevents the conversion by instantiating the SimpleTypeHandler
with properties length and name (whether we need to have the name property
is governed by the config flags, so is statically determined)

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/microsoft/chakracore/196)
<!-- Reviewable:end -->