Async function parsing compliance
This PR includes three changes to bring async functions up to spec in regard to syntax rules.

1. Fix missing parse error when an async function contains "use strict" and has a non-simple parameter list
2. Fix missing errors and incorrect behavior when local function scoped identifiers match async function parameter names (var, let, const, function, and class)
3. Fix missing parse error when method syntax functions contain duplicate parameters (this was a general bug but was discovered via @bterlson's async test262 tests)
  - I also made a related fix for #280

After fixing 1 it was bugging me that PnFnc's field for the parameter nodes was named pnodeArgs so I changed it to pnodeParams in a second commit.

Refer to the individual commits for more details about each change.