Enabling default arguments by default.
This changelist enables the default arguments scenario which does not
require split scope by default. For split scope scenarios a new flag is
added which is disabled by default. Also if a split scope required
scenario occurs in a normal mode a syntax error is thrown. That includes
scenarios where,
1. A function definition occurs in the formal parameters which refer to
one of the formals
2. Eval occurs in formals.

Fixed couple of bugs. Moved the split scope test cases to a separate file.