Improve JIT code page allocator virtual address use and reduced standing free page.
Change JD and ChakraDiag to work on the shared CodePageAllocators

Fix JD to separate disabled page (decommit but can't be reused) from the unused pages (commit and not used)

[Core Commit Message]
[MERGE #368] Improve JIT code page allocator virtual address use and reduced standing free page.

Merge pull request #368 from curtisman:page
Share CodePageAllocator within a thread by moving it into ThreadContext.
This reduce the amount of virtual address we reserved, and tighter usage of the prereserved segment.

Also commit only page requested instead of the whole segment if the maxFreePageCount will be exceeded.
When we don't have a lot of page used within a page allocator, we waste about ~100K per page allocator,
because we commit the whole segment (128K), and only use a couple of page. So in general, we wastes
about 500K-1M per thread depending on how many script context is in the thread context.

Refactor the code that manages HeapPageAllocators in CustomHeap into CodePageAllocator.
Rationalize the locking and IsNativeAddress checks
Reimplement IsInRange as IsInHeap to work the heap list instead of the page allocator, instead of
relying on the page allocator, which is now shared.