Full change due to Parse API got changed
Also to add debugger unittest to validate the parser time binding bug.

[Core Commit Message]
[MERGE #392] Removed the ParseType Reparse from the parser.

Merge pull request #392 from akroshg:parsebind2
Essentially it is used for *not* doing the parse time binding. However now we have moved to the parse-time binding for the debugger we don't need this differentiation anymore. Once I did that there is no need to ParseType_Reparse. So I have removed the whole usage altogether.