Move nested function array from the end of FunctionBody to seperate structure
Move nested function array from the end of FunctionBody to seperate
structure

Nested function array was allocated as a plus size at the end of
FunctionBody. FunctionBody is a fairly big structure, with random amount
of nested function at the end, FunctionBodys use different bucket from the
recycler and can cause memory fragmentation.
this change also remove field m_derivedSize which is not necessory any
more(it was there for locating nested function array), and m_nestedCount
will be moved to the new structure, it does not use memory for those
function does not have nested functions