Fix SIMD alignment AV on Dev12 build.
This fix separates the two SIMDValue structs.
Fix AV on Dev12 build. For SIMD value, we use a generic value (_SIMDValue) and an IA specific value with an intrinsic variable (_x86_SIMDValue). Initially, we had the IA specific one include the _SIMDValue struct. This causes the compiler to copy-prop pointers to _SIMDValue to _x86_SIMDValue, even though alignment is not guaranteed. For this particular case, it was when passing a parameter by reference to a SIMD op handler. A pointer to the stack slot in the caller holding the _SIMDValue is passed, it gets copied inside the hanlder to a _x86_SIMDValue. The compiler decides to eliminate the copy, and reference the pointer to the stack directly which causes hardware exception if it happens not to be aligned.