Enabling JIT for split scope scenarios in default parameters
The default parameter split scope introduced two new opcodes.
BeginBodyScope marks the end of param scope and the start of body scope.
In split scope we have to have separate scope for both param and body. So
a new scope slot is created for the body scope on this opcode.

At the beginning of the body the initial values of the symbols
corresponding to the params should get their final value from the
param scope symbols. To load the old value from the old param scope slots
LdParamSlot opcode is used.