Check if ScriptContext is actually closed instead of just marked for close in ThreadContext destructor
Check if ScriptContext is actually closed instead of just marked for close
in ThreadContext destructor so that Debugmanager can be released properly
on call to ScriptContext internal close.
Don't delete debug manager in ThreadContext destructor and let ScriptContext
delete it on close because it is created on ScriptContext constructor and
is ref counterd.