Debugger support for default parameters
I have also added 'dt' as a acronym for expanding to -debuglaunch -targeted. I have added this just for avoid typing that combination all the time when we write unittest

[Core Commit Message]
[MERGE #497] Debugger support for default parameters

Merge pull request #497 from akroshg:default_dbg
When we stop at formals (params) location we need to show locals, visible
only at formals scope. For that purpose DebuggerScope is used (with a new
type DiagParamScope). That scope will contain the bytecode range and symbols. With
the help of that scope we will filter variables when we are halted at
formals location.