Makes `await` a FutureReservedWord in Modules only
Fixes #507

The ES6 spec states that `await` is a _FutureReservedWord_ but then
refines this and says only when the goal symbol of the syntactic grammar
is _Module_.  Meaning that `await` cannot be used as an identifier in
ES6 modules.  The word is reserved to allow top-level await in modules
as a potential future feature addition to the language.