Fix #263: `let let` and `const let` declarations are supposed to be a syntax error
I was finally able to get a chance to finish this. I added a check to the parser that makes sure an identifier used in a `let` or `const` declaration is not `let` itself (see the links to the spec in issue #263 ). The included tests check a few of the different ways that an identifier may be used in this manner. Thanks!