Specialize pow when exponent is small integer
The current CRT in Windows 10 specializes pow(double, int) by breaking the result in to product with power of 2, then aggregation the sub-results. This leads to impricision after many float multiply operations so it was just removed from CRT which will regress Chakra performance when the CRT fix reaches public. This change does the same optimization for smaller integer which improves imaging-darkrom in kraken by 11%.