Promise.js can potentially throw an error when a promise completion handler throws an error which is not an object
This is due to the code in Promise.js being strict mode. When the
completion handler throws, we catch the error object and attempt to add
the WinRT error specific fields to it. We do this to make the stack trace
clearer. However, if the completion handler throws a non-object we end up
throwing in strict mode.

Simple fix just checks for the thrown error to be an object before
assigning properties on it. Since this error is from user code, it can
have side effects and do basically anything so surround access to it with
a try/catch just to make sure a pathological object doesn't end up
throwing out of library code.