OS:6523257: Enable CollectGarbage functionality for WWA and Webview.
During testing, I temporarily introduced some output into the function after it would have done nothing and returned if it were disabled.

The function now performs GC for WWA and Webview host types (2 and 3 respectively), and does nothing for HostType 1 (Browser). Default is HostType:1 (Browser). Applies to both internal test harness and hosting. ch.exe always acts as a Browser host.

As before, function is always present on the GlobalObject for legacy reasons (in all binaries and hosts), even if it is set to do nothing.