Default parameters pid ref stack issue
Default parameters pid ref stack issue

This is a fix for an assert happening when default param is referencing a
variable from outer scope and there is a function definition with the same
name in a block in the function body.

When we have a function definition in a block we add an additional var
declaration node for the same in the function body for backward
compatibility. When param scope also has a reference to the same name
we were adding this reference to the body scope in the wrong order.