Fix for issue when marking a function as eval caller in defer and non-defer cases
Fix for issue when marking a function as eval caller in defer and non-defer cases

In defer parse case when parsing the argument list to a call we were
not returing the argument count. We mark the current function as
CallsEval only if the eval has at least one argument in the call. This
count check was missing in the defer parse case.