remove special handling for atan2 from Js::Math::atan2
Using different methods for calculating atan2 was causing differences in result between JIT and interpreter for some inputs. Because it was unclear that the interpreter special handling of certain values is necessary, I removed the special cases and simply call the C++ library functions (as we do with all the other Math builtins).