Change runjsrttests.cmd to use temp dir instead of build dir
Stop polluting build output with unittest\jsrt\scripts.

Fixes VSO 5986161

Do to limited nature of `LoadScriptFile()` the test scripts that the
jsrt unittest binaries use need to be located within the same directory
that the unittest binaries are when running the tests.

The scripts were copied to the build output directory where the test
binaries were dropped to achieve this originally.  However this meant
that all our build archives have the *.js files scatterd in the build
output directories.

To clean that up this change uses a temp directory where it copies the
test binaries and scripts to for execution.  This way the test scripts
are not part of the build output that gets copied to the drop share.