module bugfix
fix debugger break in dynamic attach scenario
When debugger is dynamically attached, we reparse the tree and
regenerate the byte code. During the time we might need to use
the auxilary data. Disable the early release of auxilary data
for now and we'll have a better memory management when we have
full debugger support.
Also fixed a bug when the host pass in nullptr source indicating
download failure.