Remove the registering of scriptContexts those have inlineCaches
ThreadContext uses different list to keep tracking of those scriptContexts which have inlineCaches, when adding to this list it can OOM and then causes some of the inlineCaches not be able to cleared up.
Changing the design to not using this kind of registering mechanism, instead just use the flags on scriptContext to determine if there's inlineCaches need to be cleared or not. with this change there should be less work to do while allocating inlineCache, but longer list to walk while shutting down. usually we do not have many scriptContexts (my guess is in most web sites we have less than 100 scriptContexts) so it should not slow down the shutdown process.