Fix bug with recursive inlining
We do recursive inlining only when all the callsites in the function are recursive. But, the checks weren't taking into account the presence of getters and setters which are not exposed as "call" sites. This leads to data about a getter/setter present in the inliner but not in the inlinee version of the recursive function. This causes us to assert when we try to get data for the getter/setter when trying to inline it in the inlinee.