Platform Agnostic Interface for Unicode processing
This commit introduces a framework for dealing with platform specific operations
in a platform agnostic way. This change introduces the UnicodeText API in the
platform agnostic interface layer. The API replaces the following:
- NormalizeString/IsNormalizedString
- GetStringTypeW
- WindowsGlobalization::IUnicodeCharactersStatics
- LCMapStringW

In general, the APIs serve to abstract out platform specific behavior. Application specific logic remains in the calling layer.

For following new APIs are introduced in the UnicodeText namespace:
- NormalizeString - Normalizes a string according to Unicode 4.0 TR#15
- IsNormalizedString - Checks if a string is normalized according to Unicode 4.0 TR#15
- IsExternalUnicodeLibraryAvailable - Checks if a library is being used for full Unicode compat
- IsWhitespace
- IsIdStart
- IsIdContinue
- GetGeneralCategoryClass - returns a mapping from the specified character's General Category class as described in the Unicode Standard
- ChangeStringLinguisticCase - changes the case of a string based on linguistic rules using the users current locale
- GetLegacyCharacterClassificationType - Classifies the Unicode codepoint using pre-Unicode 4.0 rules
- GetLegacyCharacterTypeFlags - Gets Unicode character flags using pre-Unicode 4.0 spec

Additionally, did the following cleanup:
- Cleaned up use of ENABLE_GLOBALIZATION macro
- Removed some ES3 compat code

The linux implementations using ICU are not yet provided. They'll be provided in a subsequent checkin.