TypedArray construction from array with user-defined %ArrayIteratorPrototype%.next
TypedArray constructor uses the fast path for array on an array even if %ArrayIteratorPrototype%.next
is changed by user code. Correct behavior is to treat it as an iterable object and invoke
TypedArray(object).
Fix by adding extra conditions (essentially testing if %ArrayIteratorPrototype%.next is the built-in version)
under which TypedArray(object) will be used.
Tweak unit test to cover this case.