update interpreter to use SSE2 div when possible
In investigating a difference between JIT and interpreter, I found that interpreter is using x87 for dividing, which is significantly slower than sse2 div, and may not give same result.
To help bring consistency between JIT and interpreter, and to benefit from SSE2, I have added a path to do SSE2 div in the interpreter where possible.

With this, I'm seeing a couple small but consistent perf wins on x86: 3% improvement in Kraken crypto-aes, and 100% (1 ms to 0ms) in sunspider crypto-sha1.

While I was here, I added SSE2 multiply as well, which improves sunspider math-partial-sums by 2ms, giving an overall win of 2.5% on sunspider on my machine.