Conversion from uint32 to int32 should be marked lossy
If one of the values in a binary operation is a float constant with a value that fits in a uint32 but not an int32, and the instruction can ignore int overflow, the source value for the purposes of int specialization would have been changed to an int constant value by ignoring overflow. But, the conversion is still lossy.