ES6 Spread over zealous optimization for Arrays ignores user changes to %ArrayIteratorPrototype%.next property
This is a similar bug that was reported in #608 and fixed by @suwc for TypedArray in #635.

The core problem is that we do not check for changes to %ArrayIteratorPrototype%.next that could have side effects when optimizing spread item copying. Using a function Suwei added, we attempt to get the value without invoking implicit calls, e.g. a getter. In this case we will fall back to calling the iterator for each element.

I moved out the ArrayIteratorPrototypeHasUserDefinedNext function from TypedArray to JavascriptLibrary since we now use it in JavascriptOperators.