Profiler for call with flags does not check the call flags to see if the call is a new call
This only really happens when we are doing a super call in a derived class
constructor. That call is using ProfiledCallIFlags opcode to pass the new
flag and new target flags in the call info. When we get to the helper,
OP_ProfileCallCommon, we need to check the flags to see if we have the new
flag and pass that to RecordCallSiteInfo.

Previously we ignored the flags and always passed false for
isConstructorCall. This could lead to an assert as in the attached test
case.