MSFT: 6994835 Tweetium crashes after every launch
When we enable CFG, we preserve two more registers, r4 and r5, in calls into projection, mostly
delegate calls. In the assembly thunk that sets up the callframe for Delegate::Invoke (and others),
we need to preserve a continuous region of stack+{r0-r3} such that parameters got spread between
registers and normal stack will continue to work. Change the push/pop sequence for r4/45 to ensure that.