Fix exception handling in ScriptEngineBase::Serialize and Deserialize
We shouldn't be calling functions from a C++ catch block because stack is
unwound only after the catch block has been executed and in case of an OOM
or SO exception, a function call may lead to raising another OOM or SO
exception. Most of the places in our codebase follow the pattern of saving
the exception object in a local and taking action on that local outside
the catch. ScriptEngineBase::Deserialize and ScriptEngineBase::Serialize
were two places which were not adhering to this pattern. Fixed the code
there and also rearranged BEGIN_ENTER_SCRIPT and
BEGIN_TRANSLATE_EXCEPTION_AND_ERROROBJECT_TO_HRESULT upon Curtis'
suggestion to correctly convert exceptions to hresults.