OS:7112681 Promise callback handler code is not marked as being handled
If that code throws, the exception gets propagated back to the host via ReportError which prints it into the console in Edge. We can fix this by setting a flag before calling into the callback code which indicates that the exception should be treated as handled and not handed back to the host. Also removes one dead function from JavascriptPromise.