Fix exception path when excpeption is thrown while processing jit transfer data
After generating the jitted code for a function, we pin type references and install property guards on the entryPointInfo. If either one of these throws (OOM/SO, for example), we execute an object destructor that calls a function that allocates a new entryPointInfo. This is a problem if we were throwing an OOM exception. Apart from exceptions, this function can be called when we lazily bailout - in that case, it is ok to create a new entry point info and generate jitted code for it. In fact, we have to do this, otherwise we'll keep bailing out. For OOM scenarios, we dont really have to create a new default entry point info- we could just change the entry point to CheckCodeGenThunk. If the exception was caught and the function was called again, it'll re-use the jitted code and attempt to pin type refs and install property guards again.