Loop upper/lower bound issue
Fix the underlying loop count identification bug that was exposed by a memop issue that hit on the web. The fix was suggested by Kount: given two possibly symbols that could be the bound for the loop count, if one symbol is itself a relative bound of the other, choose the other.