Update ES6Promise_DOM unit test
This unit test broke after trident changes to the micro-task queue for
promise. We need to enforce an order of the promises in the test case
since they may be completed out-of-order more frequently with the more
granular queue. That cleanup ended-up breaking console runs of the test,
though. Fix here simplifies logic to avoid property lookup on undefined
object.