ShrU_A int type spec fix
The int range for a ShrU_A result was using an int32 min/max. We detect ShrU_A whose result may not be an int32 (source may be negative, shift amount may be zero), but because the value type assigned to the result is 'Likely Int', we still use an int32 range. We could bail out on non-int32 result, or we could assign an 'Any Number' value type. I've gone with the latter, since there may be no operation downstream that is affected by the range. Benchmarks are flat.