xplat: turn on and customize PAL_STDCPP_COMPAT mode
We use PAL's stdlib implementations which are incompatible with std headers.
This was a big headache as we couldn't include std headers such as
`<stdio.h>` or `<emmintrin.h>`, directly or indirectly. To work around we
used to copy some declarations directly (into palrt.h or CommonPal.h), till
it no longer worked for some reason (copied declarations dependency?).

This change attemps to turn on `PAL_STDCPP_COMPAT` mode and customize it,
so that we can use PAL and be compatible with std headers.

The primary change is very simple: turn on `PAL_STDCPP_COMPAT` mode, and
in `pal.h` enable some non-STDCPP-COMPAT customizations that we need. (e.g.,
`printf` is defined to `PAL_printf`, thus Chakra code that uses `printf`
are all redirected to call `PAL_printf`.)

With this change we are able to `#include <emmintrin.h>` and build
`Runtime/Library/MathLibrary.cpp` and a list of `Runtime/Language/Simd???`
files.