Disable obj-type spec for Init*LetFld and Init*ConstFld opcodes
The lowerer doesn't expect InitLetFld to be obj-type speced. While it is
legal to obj-type spec InitLetFld and changing the lowerer would be one
way to fix this issue, it would not be performant as we don't share types
of activation objects yet, leading to repeated Obj-type spec bailouts for
InitLetFld. Same reasoning applies to other opcodes Init*LetFld and
Init*ConstFld opcodes.