Undo defer parse mode issue with formals
In undo defer parse mode we reconstruct the pid ref stack for the formals.
With my last change to insert a new reference for each param symbol in the
body I missed making change in the undo defer parse mode. This change adds
the code to do that for both split scope and non-split scope.