PRE bookkeeping fix
When we clear out live fields on discovering that implicit calls prevent loop optimization, we leave the same syms in the loop's initialValueFieldMap, resulting in later asserts. Two things: should use the KillAllFields API to clear out the live fields BV, and that API should walk the loop hierarchy setting the allFieldsKilled bits.