fix build break for LargeObjectHeader
in largeHeapBlock I added static assert for size of LargeObjectHeader. this is under _M_X64_OR_ARM64 which is not defined in jd Project. adding the define here just instead of changing in core because _M_X64_OR_ARM64 is valid in core and can break jd in the future if someone using it in some of the header filed