MSFT: 5119021 ScriptEngineBase::ChangeTypeToVar() should not cause EnterScriptObject
DOM is tightening up enter_script calls such that more of the DOM tree manipulation would be safe without
entering script. Cleanup the two existing usage of DisableNoScriptScope, but keep the class there in case
we might need it for some temporary API disable scenario.
With this change, I basically reverse the DispatchHelper methods such IDispatch methods are assumed to be
called while out of script, and in-script calls need to call the LeaveScript version.