cross platform: fix linux branch build break
Fixes the build break introduced with #910

Reason: CMake version needs to be 3.3 for COMPILE_LANGUAGE QUERY support

Solution could be fixing the minimum CMake version to 3.3 or implementing a
similar functionality without changing the CMake requirements.

This PR prefers the second option due to lack of availability of CMake 3.3 (from
package managers)

Current aproach is sufficient enough without introducing a new dependency.

Tested separately with `ninja-build` in addition to `make`.

~~Also, replaced `-Wl,--no-undefined` to it's old value `-Wl,-undefined,error`~~

~~`-Wl,-undefined,error` is cross platform `clang` compatible alternative~~