Memset: Native arrays typecheck and missing item check
Memset with native float arrays needs to check the type of the source because we can do `nativearray[i] = var` and bailout after the fact if the source was not a float (ie: changes the type of array).

Since memset will not change the type of the array, it needs to check before hand that the source is actually a float or int and bailout if not.

Additionally, I added a check for `IsMissingItem` for native int and float arrays.