cross platform: fix stack/heap overflow (ulong)
`ulong` (unsigned long) is 4 bytes on Windows while 8 bytes on *nix.

There is/are routines within ChakraCore assume casting from `double`
to `ulong` and then reaching to second item `[1]` is safe. However,
this ends up corrupting stack/heap on *nix.

PAL has a definition for ULONG as (unsigned int) and ULONG is already
widely used. For this reason, replacing all the `ulong` to `ULONG`