Fix for an issue with split scope in dead store optimization
Storing the arg in for the parameter and the copy logic of initial value
of the param to the corresponding symbol in the body was happening in the
same block. During dead store check were were only checking the property
sym for usage. So the initial store was getting removed. This fix adds
usage check for the stack sym too.