PR 19942: WinRtToTs.exe for generating d.ts from .winmd files
This is an implementation of WinRtToTs.exe, a tool for generating a type definition file (.d.ts) from a set of Windows Metadata (.winmd) files.
It's intended to be used by the language service to dynamically generate type definitions for Universal Windows App projects.

It takes a list of *.winmd files and an output file path as arguments.

It's built on top of the ProjectionModel code from Chakra.